fix pre-commit hooks intallation
With the documentation we had, only commit-msg hooks were installed.
What is more, we had the stages assigned wrong for specific checks (by
default, each hook was assigned to each stage, so also to commit-msg).

As a result, this would result in pre-commit (in fact: commit-msg)
checks to appear as if they would PASS, even if they were supposed
to FAIL (and were failing in CI). In fact, most of the pre-commit
hooks on files were never executed.

Refer to the pre-commit config and default values:
https://pre-commit.com/#pre-commit-configyaml---top-level

Signed-off-by: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>